requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
pydantic-settings==2.7.0
python-dotenv==1.0.1
alembic==1.13.1
psycopg2-binary==2.9.9 
passlib[bcrypt]==1.7.4
pydantic[email]==2.7.1
python-jose[cryptography]==3.3.0